Move opinion model to dedicated package
