+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/opinion/OpinionManager.java
@@ -11,8 +11,12 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import pl.lodz.p.it.opinioncollector.exceptions.opinion.OpinionNotFoundException;
 import pl.lodz.p.it.opinioncollector.exceptions.opinion.OpinionOperationAccessForbiddenException;
-import pl.lodz.p.it.opinioncollector.productManagment.ProductRepository;
 import pl.lodz.p.it.opinioncollector.exceptions.products.ProductNotFoundException;
+import pl.lodz.p.it.opinioncollector.opinion.model.Advantage;
+import pl.lodz.p.it.opinioncollector.opinion.model.Disadvantage;
+import pl.lodz.p.it.opinioncollector.opinion.model.Opinion;
+import pl.lodz.p.it.opinioncollector.opinion.model.OpinionId;
+import pl.lodz.p.it.opinioncollector.productManagment.ProductRepository;
 import pl.lodz.p.it.opinioncollector.userModule.user.User;
 import pl.lodz.p.it.opinioncollector.userModule.user.UserManager;
